+++ .gitignore
@@ -35,7 +35,7 @@ MANIFEST
 pip-log.txt
 pip-delete-this-directory.txt
 
-# Unit test / coverage reports
+# Unit test_china / coverage reports
 htmlcov/
 .tox/
 .coverage
