+++ get_subway_connection.py
@@ -1,0 +1,5 @@
+import json
+
+with open ('subway_info.json') as js:
+    data = json.load(js)
+    print(data)
